// src/types/types.d.ts
interface UserResponse {
  questionId: string;
  response: any;
}
// ...
interface Question extends WithId<Document> {
  _id?: ObjectId;
  question: string;
  type: string;
  helpVid?: string;
  helpText?: string;
  data?: any[];
  isConcatenable?: boolean;
}